pyproject.toml
[tool.poetry]
name = "digit"
version = "0.1.0"
description = "Print smart sensors monthly reports using data from the ABB cloud"
authors = ["Davide Silvestri <silvestri.eng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.0.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.0.0"
requests = "^2.27.1"
djangorestframework-dataclasses = "^1.1.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^6.2.5"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
